test(ci): relocate clean self-contained unit tests into tests/unit (#1895)
Batch 1. Case-A movers (sys.path.insert only — no path re-anchor); baseline lines
relocated by path. Each verified: pre==post --co nodeid count, green under CI seeds
12345/67890/99999.

  watchdog_unit(43) proactive_audit_unit(4) self_execute(22) telegram_login_gate(5)
  whatsapp_adapter(83) access_grant_notification_unit(4, escape-hatch helper -> no
  baseline entry)

Two Attempt-3-driven deviations from the plan's static Batch-1 list (the trial is
the only reliable classifier, per AC #1):

- test_platform_default_model.py is MIXED, not a clean mover: 9 of 12 tests make a
  real httpx.post(BASE_URL) (Connection refused, no live backend). SPLIT it — the 3
  self-contained TestGetPlatformDefaultModelUnit tests (which hold all 6 sys.modules
  findings) -> tests/unit/; the 9 live feature-flags/settings-API tests stay in
  tests/ root with an allow-root-live-test marker.

- test_cb_probe_execution_close.py STAYS PUT (reverted): it is red-and-hidden, not
  passing coverage. Standalone it fails on two product paths — the #1804
  close_execution_activity await (its mock is a plain MagicMock, never updated) and a
  seed-dependent scrub_secret_and_urls ImportError. Moving it would redden the per-PR
  gate (AC #5). Fixing its drift is a separate bug; its sys.modules baseline line is
  left at the root path.
