+++ tests/unit/test_self_execute.py
