+++ tests/unit/test_whatsapp_adapter.py
